+++ api/internal/statemachine/event_handler.go
@@ -7,7 +7,44 @@ import (
 	"time"
 )
 
-// eventHandler is a struct that holds the event handler function and its timeout.
+var (
+	_ EventHandler = &eventHandler{}
+)
+
+// EventHandler is an interface for handling state transition events in the state machine.
+type EventHandler interface {
+	// TriggerHandler triggers the event handler for a state transition.
+	TriggerHandler(ctx context.Context, fromState, toState State) error
+}
+
+// EventHandlerFunc is a function that handles state transition events. Used to report state changes.
+type EventHandlerFunc func(ctx context.Context, fromState, toState State)
+
+// EventHandlerOption is a configurable state machine option.
+type EventHandlerOption func(*eventHandler)
+
+// WithHandlerTimeout sets the timeout for the event handler to complete.
+func WithHandlerTimeout(timeout time.Duration) EventHandlerOption {
+	return func(eh *eventHandler) {
+		eh.timeout = timeout
+	}
+}
+
+// NewEventHandler creates a new event handler with the provided function and options.
+func NewEventHandler(handler EventHandlerFunc, options ...EventHandlerOption) EventHandler {
+	eh := &eventHandler{
+		handler: handler,
+		timeout: 5 * time.Second, // Default timeout
+	}
+
+	for _, option := range options {
+		option(eh)
+	}
+
+	return eh
+}
+
+// eventHandler is a struct that implements the EventHandler interface. It contains a handler function that is called when a state transition occurs, and it supports a timeout for the handler to complete.
 type eventHandler struct {
 	handler EventHandlerFunc
 	timeout time.Duration // Timeout for the handler to complete, default is 5 seconds
@@ -26,9 +63,9 @@ func (eh *eventHandler) TriggerHandler(ctx context.Context, fromState, toState S
 				err := fmt.Errorf("event handler panic from %s to %s: %v: %s\n", fromState, toState, r, debug.Stack())
 				done <- err
 			}
+			close(done)
 		}()
 		eh.handler(ctx, fromState, toState)
-		close(done)
 	}()
 
 	select {
